Fix OOD z values that were making no transform baselines perform poorly
Because the z values during training are in a small range (due to collecting data in one go without moving the set up), z values at run time are normalized to be very out of distribution (supposed to be [0,1] but can get values of 100) due to testing on a different table and after moving the lab. While this can demonstrate the value of learning our invariant transform since it (should) ignore z, this change is not intended since it makes all the baselines unusable.